+++ Knihovna/CHMU-HLAVICKA.js
@@ -1,0 +1,182 @@
+//Verze 31
+
+// Hlavička HTML stránky
+resultText += '<HTML>';
+resultText += '<HEAD>';
+    resultText += '<META charset="utf-8"/>';
+    resultText += '<TITLE>' + vystraha.id + '</TITLE>';
+
+    #import "CHMU-STYL";
+
+resultText += '</HEAD>';
+resultText += '<BODY>';
+
+var found = false;
+if (vystraha.ucel == 'Actual') {
+    // Dohledáme, zda máme alespoň jeden jev, který není OUTLOOK
+    for (var k = 0; k < krajList.length && found == false; k++)
+    {
+        for (var i = 0; i < krajList[k].info.length && found == false; i++)
+        {
+            info = krajList[k].info[i];
+
+            if (info.jev_kod && info.jev_kod != 'OUTLOOK')
+            {
+                found = true;
+            }
+        }
+
+        // Výstrahy pro okres
+        for (var o = 0; o < krajList[k].okresList.length && found == false; o++)
+        {
+            for (var i = 0; i < krajList[k].okresList[o].info.length; i++)
+            {
+                info = krajList[k].okresList[o].info[i];
+
+                if (info.jev_kod && info.jev_kod != 'OUTLOOK')
+                {
+                    found = true;
+                }
+            }
+
+            // Výstrahy pro orp
+            for (var ol = 0; ol < krajList[k].okresList[o].orpList.length && found == false; ol++)
+            {
+                for (var i = 0; i < krajList[k].okresList[o].orpList[ol].info.length; i++)
+                {
+                    info = krajList[k].okresList[o].orpList[ol].info[i];
+
+                    if (info.jev_kod && info.jev_kod != 'OUTLOOK')
+                    {
+                        found = true;
+                    }
+                }
+            }
+        }
+    }
+}
+
+// Zjistíme zda je někde příznak HPPS, SIVS nebo SVRS
+var hpps = false;
+var sivs = false;
+var svrs = false;
+
+// Výstrahy pro kraj
+for (var k = 0; k < krajList.length && (hpps == false || sivs == false || svrs == false); k++)
+{
+    for (var i = 0; i < krajList[k].info.length && (hpps == false || sivs == false || svrs == false); i++)
+    {
+        info = krajList[k].info[i];
+
+        if (info.HPPS && info.HPPS == '1')
+        {
+            hpps = true;
+        }
+
+        if (info.SIVS && info.SIVS == '1')
+        {
+            sivs = true;
+        }
+
+        if (info.SVRS && info.SVRS == '1')
+        {
+            svrs = true;
+        }
+    }
+
+    // Výstrahy pro okres
+    for (var o = 0; o < krajList[k].okresList.length && (hpps == false || sivs == false || svrs == false); o++)
+    {
+        for (var i = 0; i < krajList[k].okresList[o].info.length; i++)
+        {
+            info = krajList[k].okresList[o].info[i];
+
+            if (info.HPPS && info.HPPS == '1')
+            {
+                hpps = true;
+            }
+
+            if (info.SIVS && info.SIVS == '1')
+            {
+                sivs = true;
+            }
+
+            if (info.SVRS && info.SVRS == '1')
+            {
+                svrs = true;
+            }
+        }
+
+        // Výstrahy pro orp
+        for (var ol = 0; ol < krajList[k].okresList[o].orpList.length && (hpps == false || sivs == false || svrs == false); ol++)
+        {
+            for (var i = 0; i < krajList[k].okresList[o].orpList[ol].info.length; i++)
+            {
+                info = krajList[k].okresList[o].orpList[ol].info[i];
+
+                if (info.HPPS && info.HPPS == '1')
+                {
+                    hpps = true;
+                }
+
+                if (info.SIVS && info.SIVS == '1')
+                {
+                    sivs = true;
+                }
+
+                if (info.SVRS && info.SVRS == '1')
+                {
+                    svrs = true;
+                }
+            }
+        }
+    }
+}
+
+// Text v těle
+switch (vystraha.ucel) {
+    case 'Exercise' :
+    case 'System' :
+    case 'Test' :
+        header = 'ÚČELOVÁ INFORMACE ČHMÚ – TESTOVACÍ ZPRÁVA';
+        if (svrs && !sivs && !hpps) {
+            header += '<br/>SMOGOVÝ VAROVNÝ A REGULAČNÍ SYSTÉM'
+        }
+        if (sivs && !hpps) {
+            header += '<br/>SYSTÉM INTEGROVANÉ VÝSTRAŽNÉ SLUŽBY';
+        }
+        if (hpps) {
+            header += '<br/>PŘEDPOVĚDNÍ POVODŇOVÁ SLUŽBA ČHMÚ';
+        }
+    break;
+    case 'Actual' :
+        header = 'VÝSTRAHA ČHMÚ';
+        if (svrs && !sivs && !hpps) {
+            header = 'ZPRÁVA SMOGOVÉHO VAROVNÉHO A REGULAČNÍHO SYSTÉMU'
+        }
+        if (sivs && !hpps) {
+            header += '<br/>SYSTÉM INTEGROVANÉ VÝSTRAŽNÉ SLUŽBY';
+        }
+        if (hpps) {
+            header += '<br/>VÝSTRAHA PŘEDPOVĚDNÍ POVODŇOVÉ SLUŽBY ČHMÚ';
+        }
+        if (!found) {
+            header = 'INFORMAČNÍ ZPRÁVA ČHMÚ';
+        }
+    break;
+
+}
+
+resultText += '<div class="header">' + header + '</div>';
+resultText += '<br/>Zpráva č. ' + vystraha.id.substring(vystraha.id.length - 6);
+resultText += '<br/>Odesláno: ' + ZobrazDatum(vystraha.dc_odeslano, 'long');
+
+if (vystraha.reference)
+{
+    var referenceSplit = vystraha.reference.split(',');
+    resultText += '<br/>Zpráva aktualizuje předchozí zprávu č. ' + referenceSplit[1].substring(referenceSplit[1].length - 6)
+        + ' vydanou ' + referenceSplit[2].substring(8, 10) + '.' + referenceSplit[2].substring(5, 7) + '.' + referenceSplit[2].substring(0, 4)
+        + ' v ' + referenceSplit[2].substring(11, 19) + ' hodin';
+}
+
+resultText += vystraha.poznamka ? '<br/>Poznámka: ' + vystraha.poznamka : '';