+++ Knihovna/CHMU-MAIL-2.js
@@ -1,4 +1,4 @@
-//Verze 30
+//Verze 31
 
 #import "CHMU-CISELNIK";
 #import "CHMU-DATUMY";
@@ -88,6 +88,15 @@ function PrepareInfo(vystraha) {
     }
     
     infoList = infoListFilter;
+
+    infoList = infoList.sort(function (a, b) {
+        var start1 = parseFloat(a.dc_zacatek);
+        var start2 = parseFloat(b.dc_zacatek);
+
+        if (start1 < start2) return -1;
+        if (start1 > start2) return 1;
+        return 0;
+    });
 
     return infoList;
 }
@@ -214,8 +223,7 @@ function PrepareKraje(orp, infoList) {
     return krajList;
 }
 
-function GetWarningColor(info)
-{
+function GetWarningColor(info) {
     // Barva podle závažnosti
     var color = '';
 
@@ -490,18 +498,6 @@ function PrintInfo(info, ref_info) {
 
     return resultText;
 }
-
-// Odstranění duplicitních výskytů kódů krajů
-function removeDuplicates(arr) {
-    var unique_array = []
-    for(var i = 0;i < arr.length; i++){
-        if(unique_array.indexOf(arr[i]) == -1){
-            unique_array.push(arr[i])
-        }
-    }
-    return unique_array
-}
-
 
 function Remove(arr, item) {
     for (var i = arr.length; i--;) {
@@ -1178,8 +1174,8 @@ function JevUzemi(info) {
     }
 
     // Česká republika
-    if (uzemiKraje.indexOf('19') != -1 && uzemiKraje.indexOf('27') != -1 && uzemiKraje.indexOf('35') != -1 && uzemiKraje.indexOf('43') != -1 && uzemiKraje.indexOf('51') != -1 && uzemiKraje.indexOf('60') != -1 && uzemiKraje.indexOf('78') != -1 && uzemiKraje.indexOf('86') != -1 && uzemiKraje.indexOf('94') != -1 && uzemiKraje.indexOf('108') != -1 && uzemiKraje.indexOf('116') != -1 && uzemiKraje.indexOf('124') != -1 && uzemiKraje.indexOf('141') != -1 && uzemiKraje.indexOf('132') != -1) {
-        resultText = 'Česká republika';
+    if (resultText == '<b><u>Hlavní město Praha</u></b>, <b><u>Středočeský kraj</u></b>, <b><u>Jihočeský kraj</u></b>, <b><u>Plzeňský kraj</u></b>, <b><u>Karlovarský kraj</u></b>, <b><u>Ústecký kraj</u></b>, <b><u>Liberecký kraj</u></b>, <b><u>Královéhradecký kraj</u></b>, <b><u>Pardubický kraj</u></b>, <b><u>Kraj Vysočina</u></b>, <b><u>Jihomoravský kraj</u></b>, <b><u>Olomoucký kraj</u></b>, <b><u>Moravskoslezský kraj</u></b>, <b><u>Zlínský kraj</u></b>\n') {
+        resultText = '<b><u>ČESKÁ REPUBLIKA</u></b>';
     }
 
     resultText += '|' + uzemiList.length;
@@ -1241,189 +1237,11 @@ if (typeof(ref_vystraha) != 'undefined' && ref_vystraha.info && ref_vystraha.inf
     ref_krajList = PrepareKraje(orp, ref_infoList);
 }
 
-// Hlavička HTML stránky
-resultText += '<HTML>';
-resultText += '<HEAD>';
-    resultText += '<META charset="utf-8"/>';
-    resultText += '<TITLE>' + vystraha.id + '</TITLE>';
-
-    #import "CHMU-STYL";
-
-resultText += '</HEAD>';
-resultText += '<BODY>';
-
-var found = false;
-if (vystraha.ucel == 'Actual') {
-    // Dohledáme, zda máme alespoň jeden jev, který není OUTLOOK
-    for (var k = 0; k < krajList.length && found == false; k++)
-    {
-        for (var i = 0; i < krajList[k].info.length && found == false; i++)
-        {
-            info = krajList[k].info[i];
-
-            if (info.jev_kod && info.jev_kod != "OUTLOOK")
-            {
-                found = true;
-            }
-        }
-
-        // Výstrahy pro okres
-        for (var o = 0; o < krajList[k].okresList.length && found == false; o++)
-        {
-            for (var i = 0; i < krajList[k].okresList[o].info.length; i++)
-            {
-                info = krajList[k].okresList[o].info[i];
-
-                if (info.jev_kod && info.jev_kod != "OUTLOOK")
-                {
-                    found = true;
-                }
-            }
-
-            // Výstrahy pro orp
-            for (var ol = 0; ol < krajList[k].okresList[o].orpList.length && found == false; ol++)
-            {
-                for (var i = 0; i < krajList[k].okresList[o].orpList[ol].info.length; i++)
-                {
-                    info = krajList[k].okresList[o].orpList[ol].info[i];
-
-                    if (info.jev_kod && info.jev_kod != "OUTLOOK")
-                    {
-                        found = true;
-                    }
-                }
-            }
-        }
-    }
-}
-
-// Zjistíme zda je někde příznak HPPS, SIVS nebo SVRS
-var hpps = false;
-var sivs = false;
-var svrs = false;
-
-// Výstrahy pro kraj
-for (var k = 0; k < krajList.length && (hpps == false || sivs == false || svrs == false); k++)
-{
-    for (var i = 0; i < krajList[k].info.length && (hpps == false || sivs == false || svrs == false); i++)
-    {
-        info = krajList[k].info[i];
-
-        if (info.HPPS && info.HPPS == "1")
-        {
-            hpps = true;
-        }
-
-        if (info.SIVS && info.SIVS == "1")
-        {
-            sivs = true;
-        }
-
-        if (info.SVRS && info.SVRS == "1")
-        {
-            svrs = true;
-        }
-    }
-
-    // Výstrahy pro okres
-    for (var o = 0; o < krajList[k].okresList.length && (hpps == false || sivs == false || svrs == false); o++)
-    {
-        for (var i = 0; i < krajList[k].okresList[o].info.length; i++)
-        {
-            info = krajList[k].okresList[o].info[i];
-
-            if (info.HPPS && info.HPPS == "1")
-            {
-                hpps = true;
-            }
-
-            if (info.SIVS && info.SIVS == "1")
-            {
-                sivs = true;
-            }
-
-            if (info.SVRS && info.SVRS == "1")
-            {
-                svrs = true;
-            }
-        }
-
-        // Výstrahy pro orp
-        for (var ol = 0; ol < krajList[k].okresList[o].orpList.length && (hpps == false || sivs == false || svrs == false); ol++)
-        {
-            for (var i = 0; i < krajList[k].okresList[o].orpList[ol].info.length; i++)
-            {
-                info = krajList[k].okresList[o].orpList[ol].info[i];
-
-                if (info.HPPS && info.HPPS == "1")
-                {
-                    hpps = true;
-                }
-
-                if (info.SIVS && info.SIVS == "1")
-                {
-                    sivs = true;
-                }
-
-                if (info.SVRS && info.SVRS == "1")
-                {
-                    svrs = true;
-                }
-            }
-        }
-    }
-}
-
-// Text v těle
-switch (vystraha.ucel) {
-    case 'Exercise' :
-    case 'System' :
-    case 'Test' :
-        header = "ÚČELOVÁ INFORMACE ČHMÚ - TESTOVACÍ ZPRÁVA";
-        if (svrs && !sivs && !hpps) {
-            header += '<br/>SMOGOVÝ VAROVNÝ A REGULAČNÍ SYSTÉM'
-        }
-        if (sivs && !hpps) {
-            header += '<br/>SYSTÉM INTEGROVANÉ VÝSTRAŽNÉ SLUŽBY';
-        }
-        if (hpps) {
-            header += '<br/>PŘEDPOVĚDNÍ POVODŇOVÁ SLUŽBA ČHMÚ';
-        }
-    break;
-    case 'Actual' :
-        header = "VÝSTRAHA ČHMÚ";
-        if (svrs && !sivs && !hpps) {
-            header = 'ZPRÁVA SMOGOVÉHO VAROVNÉHO A REGULAČNÍHO SYSTÉMU'
-        }
-        if (sivs && !hpps) {
-            header += '<br/>SYSTÉM INTEGROVANÉ VÝSTRAŽNÉ SLUŽBY';
-        }
-        if (hpps) {
-            header += '<br/>VÝSTRAHA PŘEDPOVĚDNÍ POVODŇOVÉ SLUŽBY ČHMÚ';
-        }
-        if (!found) {
-            header = "INFORMAČNÍ ZPRÁVA ČHMÚ";
-        }
-    break;
-}
-
-resultText += '<div class="header">' + header + '</div>';
-resultText += '<br/>Zpráva č. ' + vystraha.id.substring(vystraha.id.length - 6);
-resultText += '<br/>Odesláno: ' + ZobrazDatum(vystraha.dc_odeslano, 'long');
-
-if (vystraha.reference)
-{
-    var referenceSplit = vystraha.reference.split(',');
-    resultText += '<br/>Zpráva aktualizuje předchozí zprávu č. ' + referenceSplit[1].substring(referenceSplit[1].length - 6)
-        + ' vydanou ' + referenceSplit[2].substring(8, 10) + '.' + referenceSplit[2].substring(5, 7) + '.' + referenceSplit[2].substring(0, 4)
-        + ' v ' + referenceSplit[2].substring(11, 19) + ' hodin';
-}
-
-resultText += vystraha.poznamka ? '<br/>Poznámka: ' + vystraha.poznamka : '';
+#import "CHMU-HLAVICKA";
 
 resultText += '<br/>Územní platnost: ';
 if (omezitNaKraj == '-1') {
-    resultText += "Česká republika";
+    resultText += 'Česká republika';
 } else {
     resultText += KRAJE_NAZVY[omezitNaKraj];
 }